PDBS: add package check for getMaximumDataBlockSize()
This method is only useful to the package that writes data to the block
so this applies the same security check as the write() method.

getDataBlockSize()'s security check was only relaxed to use a permission
as it is used by other system apps to determine whether the block
contains data and hence is in use (see ag/593371).

Fix: 62348164
Test: add a Google account to the device causing gmscore to call the
method
Change-Id: I7e71ea236899bf2e4df7cb26ae285b414e9e6565