Use configuration delta to determine if still in orientation change.
We currently track this change in the WindowState with a variable
called mOrientationChanging. This value is set and cleared in a
variety of classes. It is possible for this value to be cleared
before processing an orientation changing frame. This can lead to us
prematurely unfreezing.

Since the orientation is present in the configuration, we can use the
orientation delta between the last and current frame's configuration
to determine whether we are in the middle of an orientation change.
The existing signal has been moved behind a setter/getter where the
latter is combined with the configuration state.

This reverts commit ee50e9a83f6c7d1fa34a734b662445a26745b96f. It
addresses issues raised in b/63380251 where the orientation delta was
being considered for WindowStates that were not visible.

Bug: 62846907
Test: go/wm-smoke
Change-Id: I574efb8f9a458cb295f2ef14e10dbb91e8875bf4