Ignore "mLastHidden" state when deciding whether to turnScreenOn.
In the case of showWhileLocked, turnScreenOn, windows, we may not
hide them or destroy their surface when turning the screen off. In
this case mLastHidden will never be set to true, and we will never
enter the showSurfaceRobustlyLocked path. However this is the path where
we consume turnScreenOn, and so this was preventing such a window
from being able to turn the screen on a second time.

Bug: 62913759
Test: Manual from bug. go/wm-smoke. "At some form at a later date"
Change-Id: Ia72d702e175b82c5f0e628010177f02bf87abf84