Report the display is off when the hardware blanks itself.
Some hardware will blank itself coming out of doze to avoid artifacts
from being shown. In order to perform a graceful transition from doze to
on, tell window manager policy that the display turned off and then
turned on without actually telling the display to enter the off state,
in order to keep the screen on latency low.

Bug: 63531607
Test: manual
Change-Id: I4362905e47d5fb47b639c8701c91d69b50aa1232