Avoid entering orientation changing state when seamlessly rotating.
It used to be that we only used mOrientationChanging instead of the
configuration mismatch detection. We would explicitly avoid setting
mOrientationChanging in the case of seamless rotation. We need to
continue to return false here for seamless rotation as getOrientationChanging
is only used to track the screen freezing logic which we don't want here.
It can also trigger windows to be re-shown, causing animations and glitches
which are presumably normally covered up by the screenshot.

Bug: 63852205
Test: Manual, go/wm-smoke
Change-Id: I537fa815180bbccc3b9682a15496d3ef851df6ee